Add AVR_Periph as submodule
AVR_Periph was added as a submodule as it unites multiple AVR components
into specific use case. A first peripheral class was written to handle
motor operations.

+++ io/DigitalOutputPin.cpp
@@ -8,7 +8,7 @@
 DigitalOutputPin::DigitalOutputPin (volatile uint8_t* port, uint8_t bit)
 : m_Port(port), m_State(false)
 {
-    m_Mask = static_cast<uint8_t>(1) << bit;
+    m_Mask = 1 << bit;
     m_MaskInv = ~m_Mask;
     *(m_Port - 1) |=  m_Mask;
 }
